Create index.html
Task: 
When the page loads, a popup has to be displayed with the following question: "Do you know a bit already about Javascript?" 
When clicking on "yes" the popup close itself and the message ":)" has to be displayed in the console of the browser.

I usually always write html, css and js in separate files, but this exercise was so short that I combined js with html file. No css was needed.